Move the formatter code to a dedicated module.
